Rename the file as per review comment
