+++ frontend/src/metabase/api/index.ts
@@ -1,3 +1,3 @@
 export * from "./api";
 export * from "./api-keys";
-export * from "./native-dataset";
+export * from "./dataset";
